frame size dependency on sample rate && extra files
The driver has hardcoded dependency for period size on sample rate.
This restricts driver from being used for all sample rate and period
size combinations.

The driver also introduces an unnecessary delay in giving the first
buffer to dsp both for playback and record case. resulting in delay
in start of session.

Change-Id: I9e0eb9b426e4f77c161dc7ef72b152d081979013
Signed-off-by: Asish Bhattacharya <asishb@codeaurora.org>

Conflicts:
	sound/soc/msm/msm-pcm-afe.c